fix: cast subscriptionId to uint32 in Raffle deployment

+++ script/RaffleDeploy.s.sol
@@ -21,7 +21,7 @@ contract RaffleDeploy is Script {
             config.vrfCoordinator,
             config.gasLane,
             config.callbackGasLimit,
-            config.subscriptionId
+            uint32(config.subscriptionId)
         );
         vm.stopBroadcast();
         return (raffle, helperConfig);
